+++ Vault-Coder/src/utopia/vault/coder/model/datatype/CustomPropertyType.scala
@@ -2,6 +2,7 @@ package utopia.vault.coder.model.datatype
 
 import utopia.flow.datastructure.immutable.{ModelDeclaration, ModelValidationFailedException}
 import utopia.flow.datastructure.template
+import utopia.flow.datastructure.template.{Model, Property}
 import utopia.flow.generic.ValueUnwraps._
 import utopia.flow.generic.{FromModelFactory, StringType}
 import utopia.flow.parse.Regex
@@ -9,9 +10,11 @@ import utopia.flow.util.CollectionExtensions._
 import utopia.flow.util.StringExtensions._
 import utopia.vault.coder.model.data.Name
 import utopia.vault.coder.model.datatype.CustomPropertyType.{pluralRegex, singularRegex, valueRegex}
+import utopia.vault.coder.model.datatype.CustomPropertyType.CustomPartConversion
 import utopia.vault.coder.model.enumeration.NamingConvention.CamelCase
 import utopia.vault.coder.model.scala.code.CodePiece
 import utopia.vault.coder.model.scala.datatype.{Reference, ScalaType}
+import utopia.vault.coder.model.scala.template.ValueConvertibleType
 
 import scala.util.{Failure, Success}
 
@@ -27,7 +30,7 @@ object CustomPropertyType extends FromModelFactory[CustomPropertyType]
 	private val singularRegex = parameterIndicatorRegex + "s"
 	private val pluralRegex = parameterIndicatorRegex + "p"
 	
-	private lazy val schema = ModelDeclaration("type" -> StringType, "sql" -> StringType)
+	private lazy val schema = ModelDeclaration("type" -> StringType)
 	
 	
 	// IMPLEMENTED  -------------------------
@@ -35,37 +38,82 @@ object CustomPropertyType extends FromModelFactory[CustomPropertyType]
 	override def apply(model: template.Model[template.Property]) =
 		schema.validate(model).toTry.flatMap { model =>
 			// from_value, from_values and to_value must all exist and contain the appropriate parameter placeholder
-			Vector("from_value", "to_value", "option_from_value")
-				.findMap { propName =>
-					model(propName).string match {
-						case Some(code) =>
-							if (valueRegex.existsIn(code))
-								None
-							else
-								Some(s"$propName doesn't contain the appropriate value reference ('$valueRegex')")
-						case None => Some(s"$propName is missing")
+			ensureFunctions(model, Vector("from_value", "to_value", "option_from_value")).flatMap { _ =>
+				// There must be either "parts" -property (multi-column) or "sql" property (single-column)
+				// Plus, the parts must be parseable
+				model("parts").getVector.tryMap { v => CustomPartConversion(v.getModel) }.flatMap { parts =>
+					if (parts.nonEmpty || model.containsNonEmpty("sql")) {
+						val default: CodePiece = model("default", "default_value")
+						// Handles the single-column use-case, if necessary
+						val conversion = if (parts.isEmpty) Left(sqlTypeFrom(model, default)) else Right(parts)
+						Success(apply(ScalaType(model("type")), conversion, model("from_value"),
+							model("option_from_value"), model("to_value"), model("option_to_value"),
+							model("empty", "empty_value"), default,
+							model("prop_name", "property_name", "default_prop_name", "default_name"),
+							model("description", "doc", "desc"), model("from_value_can_fail", "yields_try", "try")))
 					}
-				} match
-			{
-				case Some(failureMessage) => Failure(new ModelValidationFailedException(failureMessage))
-				case None =>
-					val default: CodePiece = model("default")
-					val sqlDefault = model("sql_default").stringOr { default.toSql.getOrElse("") }
-					val sqlType = SqlPropertyType(model("sql"), sqlDefault,
-						model("col_suffix", "column_suffix", "suffix"),
-						indexByDefault = model("index", "is_index", "default_index", "default_indexing"))
-					Success(apply(ScalaType(model("type")), sqlType, model("from_value"),
-						model("option_from_value"), model("to_value"), model("option_to_value"), default,
-						model("prop_name", "property_name", "default_prop_name", "default_name"),
-						model("description", "doc", "desc"),
-						model("from_value_can_fail", "yields_try", "try")))
+					else
+						Failure(new ModelValidationFailedException(
+							"A custom data type must either define the 'sql' or the 'parts' -property"))
+				}
 			}
 		}
 	
 	
+	// OTHER    --------------------------
+	
+	private def ensureFunctions(model: Model[Property], functionNames: Iterable[String]) = {
+		functionNames
+			.findMap { propName =>
+				model(propName).string match {
+					case Some(code) =>
+						if (valueRegex.existsIn(code))
+							None
+						else
+							Some(s"$propName doesn't contain the appropriate value reference ('$valueRegex')")
+					case None => Some(s"$propName is missing from $model")
+				}
+			} match
+		{
+			case Some(failureMessage) => Failure(new ModelValidationFailedException(failureMessage))
+			case None => Success(())
+		}
+	}
+	
+	private def sqlTypeFrom(model: Model[Property], defaultValue: CodePiece = CodePiece.empty) = {
+		val sqlDefault = model("sql_default").stringOr { defaultValue.toSql.getOrElse("") }
+		SqlPropertyType(model("sql"), sqlDefault,
+			model("col_suffix", "column_suffix", "suffix"), model("optional", "nullable", "allows_null"),
+			model("index", "is_index", "default_index", "default_indexing"))
+	}
+	
+	
 	// NESTED   --------------------------
 	
+	object CustomPartConversion extends FromModelFactory[CustomPartConversion]
+	{
+		// NB: Type refers to the mid-state. i.e. what the type is in the dbModel construction parameter
+		private val schema = ModelDeclaration("type" -> StringType, "sql" -> StringType)
+		
+		override def apply(model: Model[Property]) =
+			schema.validate(model).toTry.flatMap { model =>
+			// extract, extract_from_option must exist and contain the appropriate parameter placeholder ($v)
+			ensureFunctions(model, Vector("extract", "extract_from_option")).map { _ =>
+				val toValueInput: CodePiece = model("to_value")
+				val emptyValue = CodePiece.fromValue(model("empty_value", "empty"))
+					.filter { _.nonEmpty }.getOrElse(CodePiece.none)
+				apply(ScalaType(model("type")), sqlTypeFrom(model), model("extract"), model("extract_from_option"),
+					toValueInput.notEmpty
+						.getOrElse { CodePiece(parameterIndicatorRegex.string, Set(Reference.valueConversions)) },
+					emptyValue)
+				}
+			}
+	}
 	
+	case class CustomPartConversion(midType: ScalaType, sqlType: SqlPropertyType, extractPart: CodePiece,
+	                                extractPartFromOption: CodePiece,
+	                                partToValue: CodePiece = CodePiece(parameterIndicatorRegex.string, Set(Reference.valueConversions)),
+	                                emptyMidValue: CodePiece = CodePiece.none)
 }
 
 /**
@@ -73,7 +121,9 @@ object CustomPropertyType extends FromModelFactory[CustomPropertyType]
   * @author Mikko Hilpinen
   * @since 17.7.2022, v1.15.1
   * @param scalaType The wrapped scala data type
-  * @param sqlType Sql representation of this property type
+  * @param conversion Either Left: The underlying SQL data type (just the type) behind this data type
+  *                   (single-column use-case) or Right: The parts (column-representations) that form this combined
+  *                   data type (multi-column use-case)
   * @param fromValue Code that takes a Value (represented by $v) and returns an instance of this type
   * @param optionFromValue Code that takes a Value ($v) and returns an option containing an instance of this type
   * @param toValue Code that takes an instance of this type (represented by $v) and returns a Value
@@ -88,24 +138,37 @@ object CustomPropertyType extends FromModelFactory[CustomPropertyType]
   *                        replaced with a plural class name.
   * @param yieldsTryFromValue Whether fromValue code yields a Try instead of an instance of this type (default = false)
   */
-case class CustomPropertyType(scalaType: ScalaType, sqlType: SqlPropertyType, fromValue: CodePiece, optionFromValue: CodePiece,
+case class CustomPropertyType(scalaType: ScalaType, conversion: Either[SqlPropertyType, Vector[CustomPartConversion]],
+                              fromValue: CodePiece, optionFromValue: CodePiece,
                               toValue: CodePiece, optionToValue: CodePiece = CodePiece.empty,
                               emptyValue: CodePiece = CodePiece.empty,
-                              nonEmptyDefaultValue: CodePiece = CodePiece.empty,
-                              defaultPropName: Option[Name] = None, autoDescription: String = "",
-                              yieldsTryFromValue: Boolean = false)
-	extends ConcretePropertyType
+                              nonEmptyDefaultValue: CodePiece = CodePiece.empty, defaultPropName: Option[Name] = None,
+                              autoDescription: String = "", yieldsTryFromValue: Boolean = false)
+	extends PropertyType
 {
+	// ATTRIBUTES   ----------------------------
+	
+	override lazy val sqlConversions = conversion match {
+		case Left(targetType) => Vector(new OptionWrappingSqlConversion(targetType))
+		case Right(parts) => parts.map { new PartToSqlConversion(_) }
+	}
+	
+	
 	// IMPLEMENTED  ----------------------------
+	
+	override def concrete = this
+	override def optional: PropertyType = OptionWrapped
 	
 	override def defaultPropertyName =
 		defaultPropName.getOrElse { Name.interpret(scalaType.toString, CamelCase.capitalized) }
 	
-	// TODO: Add support for multi-column data-types
-	override def fromValueCode(valueCode: String, multipleValues: Boolean) = finalizeCode(fromValue, valueCode)
+	override def fromValueCode(valueCode: String) = finalizeCode(fromValue, valueCode)
+	override def fromValuesCode(valuesCode: String) =
+		fromValueCode("v").mapText { fromValue => s"$valuesCode.map { v => $fromValue }" }
 	override def toValueCode(instanceCode: String) = finalizeCode(toValue, instanceCode)
-	override def optionFromValueCode(valueCode: String) = finalizeCode(optionFromValue, valueCode)
-	override def optionToValueCode(optionCode: String) = optionToValue.notEmpty match {
+	
+	private def optionFromValueCode(valueCode: String) = finalizeCode(optionFromValue, valueCode)
+	private def optionToValueCode(optionCode: String) = optionToValue.notEmpty match {
 		// Case: Using a user-defined conversion
 		case Some(toValue) => finalizeCode(toValue, optionCode)
 		// Case: No user-defined conversion available
@@ -125,4 +188,75 @@ case class CustomPropertyType(scalaType: ScalaType, sqlType: SqlPropertyType, fr
 	
 	private def finalizeCode(userCode: CodePiece, parameterCode: String) =
 		userCode.mapText { _.replaceAll(valueRegex, parameterCode) }
+	
+	
+	// NESTED   ----------------------------
+	
+	private object OptionWrapped extends PropertyType
+	{
+		override lazy val scalaType = ScalaType.option(CustomPropertyType.this.scalaType)
+		override lazy val sqlConversions = conversion match {
+			case Left(targetType) => Vector(DirectSqlTypeConversion(OptionWrappedMidType, targetType))
+			case Right(parts) => parts.map { new PartToSqlConversion(_, extractFromOption = true) }
+		}
+		
+		override def emptyValue = CodePiece.none
+		override def nonEmptyDefaultValue = CodePiece.empty
+		
+		override def defaultPropertyName = CustomPropertyType.this.defaultPropertyName
+		
+		override def yieldsTryFromValue = false
+		
+		override def optional = this
+		override def concrete = CustomPropertyType.this
+		
+		override def toValueCode(instanceCode: String) = optionToValueCode(instanceCode)
+		override def fromValueCode(valueCode: String) = optionFromValueCode(valueCode)
+		override def fromValuesCode(valuesCode: String) =
+			fromValueCode("v").mapText { fromValue => s"$valuesCode.flatMap { v => $fromValue }" }
+		
+		override def writeDefaultDescription(className: Name, propName: Name) =
+			CustomPropertyType.this.writeDefaultDescription(className, propName)
+	}
+	
+	private object OptionWrappedMidType extends ValueConvertibleType
+	{
+		override def scalaType = OptionWrapped.scalaType
+		
+		override def emptyValue = CodePiece.none
+		
+		override def toValueCode(instanceCode: String) = optionToValueCode(instanceCode)
+	}
+	
+	private class OptionWrappingSqlConversion(targetType: SqlPropertyType) extends SqlTypeConversion
+	{
+		override def origin = CustomPropertyType.this.scalaType
+		override def intermediate = OptionWrappedMidType
+		override def target = targetType
+		
+		override def midConversion(originCode: String) = s"Some($originCode)"
+	}
+	
+	private class PartToSqlConversion(part: CustomPartConversion, extractFromOption: Boolean = false)
+		extends SqlTypeConversion
+	{
+		override def origin =
+			if (extractFromOption) OptionWrapped.scalaType else CustomPropertyType.this.scalaType
+		override def intermediate: ValueConvertibleType = IntermediateState
+		override def target = part.sqlType
+		
+		override def midConversion(originCode: String) =
+			finalizeCode(if (extractFromOption) part.extractPartFromOption else part.extractPart, originCode)
+		
+		
+		// NESTED   ------------------------
+		
+		private object IntermediateState extends ValueConvertibleType
+		{
+			override def scalaType = part.midType
+			override def emptyValue = part.emptyMidValue
+			
+			override def toValueCode(instanceCode: String) = finalizeCode(part.partToValue, instanceCode)
+		}
+	}
 }